feat: encrypt images on IPDR upload
NOTE: currently, the image's config is not encrypted (due to lack of support for config encryption in the underlying ocicrypt library). This means that we still leak the relative order of layers (though due to encryption, this reveals only layer sizes), any ENV-s set in the Dockerfile, and the default CMD/ARGs.