chore: license under GPL-3.0 (and add SPDX-License-Identifier-s)
Rationale for picking GPL-3.0: All libraries we use are MIT, BSD-3-Clause, or Apache-2.0, except the go-ethereum library which is GPL-3.0. As there is no way to link to a GPL-3.0 library other than licensing our code under GPL-3.0 as well, we have to license the whole project under the GPL-3.0 license.

(That being said, individual components like the p2p-helper and some of the non-ethereum code in pkg/ could be just as well licensed as MIT/Apache-2.0; however, the complexity of having to manage two licenses is not worth the pain at this stage. If necessarily, such relicensing of components could take place later on, but, for now, it is superfluous.)