refactor!: don't upload yamls, relocate and streamline registry code
Moved provider.HostInfo to a newly-added registry.proto, since decoding YAML received from untrusted sources is a potential security risk (even if not in the Go client, other/future clients are likely to suffer.) Coincidentially, also removed the dependency of the publisher client on provider code.

Merged pkg/registry and pkg/publisher, for consistency with the rest of the codebase as the registry only included publisher client code.

Made cmd/tpodserver automatically add in the p2p address from IPFS. Ideally, this would also add the rest of the multiaddrs the node announces publicly, but deemed that unnecessary for the time being.

Made cmd/trustedpods not require a --token-contract flag when using the registry, by allowing passing the --payment-contract for that token contract. Tho, thinking the registry should probably just store tables by payment contract address, like the rest of the code does.

Changed cmd/tpodserver to use info.contents and info.format in the config format, for consistency with pricing table configuration.

And a few other misc refactoring changes.