fix!: kvm2 create private subnets, switch to virtualbox driver
We want to allow the clusters to be reached by the client and with each other,
since by design mnikube runs each cluster in a private subnets when
using kvm driver, simply adding routes using route command does not work,
also tried port-forwarding in the host and sending requests within the vm to the host
and it didn't work, and to fix it this you need to manually edit the network
settings from outside and destroy it and restart the vm as explained here
https://github.com/kubernetes/minikube/issues/11499, which makes the tests harder to run.

A solution would be to use docker and create a loop until all kube-proxy
pods are running as needed, but it can only do 2 clusters for some
reason, the third will automatically use kvm2. moreover it behaves
differently for each run.

But that won't be needed as virtualbox driver runs all clusters within the same network.

Also registered 3 different providers, one for each cluster using their
own ipfs node. and simplified the scripts.