fix!: force containerd to pull already saved tagged images
Images were pulled from remote registries and not from the already
saved images from ipfs during pod creation. so i tried to tag the images with
the CID and replace the image url with the tag. this also fixes the case where the
client uploads from private registries, because they will be
inaccessible such as the case with the autoscaler in the test that is
stored in the local docker registry.

The problem was that containerd will always prefix the tagged images
with docker.io/ to resolve them remotely again, therefore bypassing the
saved images from ipfs. and unless its a known registry, such as gcr or
docker.io, it will always get prefixed, even localhost.

So the solution was simple,just disable pulling images completely,
therefore it will resort to the saved images with prefixing, it just
takes it as it is.